app/requirements.txt
asgiref==3.11.0
Django==5.0.10
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
whitenoise==6.6.0
psycopg[binary]==3.1.18
python-dotenv==1.0.0
django-cors-headers
djangorestframework
Pillow>=10.0.0